Split visual novel cutscenes into per-quest files like quests
- Add shared.ts with videoOnly helper
- Create mnog.ts, story.ts, mask_hunt.ts, bohrok_swarm.ts, bohrok_kal.ts, mask_of_light.ts
- Update index.ts to aggregate and export VISUAL_NOVEL_CUTSCENES
- Remove monolithic visualNovel.ts
