+++ src/data/cutscenes/bohrok_kal.ts
@@ -1,0 +1,518 @@
+import type { VisualNovelCutscene } from '../../types/Cutscenes';
+
+export const BOHROK_KAL_CUTSCENES: Record<string, VisualNovelCutscene> = {
+  bohrok_kal_reconstruction: {
+    id: 'bohrok_kal_reconstruction',
+    background: { type: 'gradient', from: '#5b3e0b', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Tahu. Come with me. There is something I must show you—away from the others.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What is it, Turaga? You look troubled.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The Nuva Symbol that appeared on the Suva... it is a sign of your transformation. Your power is now tied to it. Guard it well.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `But that is not why I called you here. I have kept a secret—one that weighs heavily on an old Turaga's conscience. Take this.`,
+      },
+      {
+        type: 'narration',
+        text: `Vakama reaches into his cloak and produces a mask unlike any Tahu has seen. Its surface shimmers, distorting the light around it as if time itself bends in its presence.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What... what is this mask?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The Kanohi Vahi. The Great Mask of Time. It is the most dangerous mask ever forged. With it, you can slow or hasten the flow of time itself.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Why give it to me?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Because a day may come when all other options fail. When that day arrives, you—and only you—must decide whether to use it. But know this: if the Vahi is damaged, time itself could unravel. All of reality could cease to exist.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I understand. I will keep it secret. And I will use it only in the direst emergency.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I pray that day never comes. Now go, Toa Nuva. Your brothers and sisters need you.`,
+      },
+    ],
+  },
+
+  bohrok_kal_scattered_aid: {
+    id: 'bohrok_kal_scattered_aid',
+    background: { type: 'gradient', from: '#1a3a4a', to: '#0d1f2d' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The Bohrok left a trail of destruction across the island. Our people need us—not as warriors, but as builders.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Agreed. Ga-Koro's platforms were damaged by the Pahrak. I should return and help the repairs.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Ko-Koro's walls need reinforcing. I will see to it.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Le-Koro was hit hard. Time to replant what the swarm tore down.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then we go to our homes. Let's give our people something to celebrate.`,
+      },
+    ],
+  },
+
+  bohrok_kal_stolen_symbols: {
+    id: 'bohrok_kal_stolen_symbols',
+    background: { type: 'gradient', from: '#1a0a2e', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'tahnok_kal',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Stand aside, Toa Nuva. You are in our way.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What are you?! You're not like the other Bohrok—`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'nuhvok_kal',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We are the Bohrok-Kal. We search for Cahdok and Gahdok, queens of the swarms. Tell us where you have hidden the Bahrag, and then stand aside. We have no wish to harm helpless foes.`,
+      },
+      {
+        type: 'narration',
+        text: `Before Tahu can respond, the six Bohrok-Kal strike simultaneously—each ripping a Nuva Symbol from its Suva shrine. Across the island, the Toa Nuva's elemental powers vanish in an instant.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `My fire—it's gone! I tried to stop one of them and it buried me under rubble like I was nothing!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I was crossing an ice bridge when my power failed. I nearly fell into the chasm.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I was highflying above Le-Wahi when my air control just... stopped. Kongu had to rescue me on a Gukko. Miles above the ground!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `A tidal wave I had been shaping collapsed on top of me. Without my power I couldn't stop it.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `These Kal are not ordinary Bohrok. They spoke to me—they said they seek the Bahrag. They want to free them.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Without our elements, how do we stop them?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We still have our masks, our weapons, and each other. We find the Kal, we take back our symbols. There is no other option.`,
+      },
+    ],
+  },
+
+  bohrok_kal_sighting: {
+    id: 'bohrok_kal_sighting',
+    background: { type: 'gradient', from: '#c2a366', to: '#1a0a2e' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Two of the Bohrok-Kal were spotted heading through Po-Wahi. If we move now, we can intercept them.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Good. We split into two groups—three to pursue the Kal, three to investigate the Bahrag's prison. We need to know if the cage still holds.`,
+      },
+      {
+        type: 'narration',
+        text: `The Toa intercept Pahrak-Kal and Kohrak-Kal in a narrow Po-Wahi canyon. The two Kal halt and turn to face them—not with aggression, but with calm certainty.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'pahrak_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `You are too late, Toa. Already, the Bohrok-Kal have learned where the Bahrag are hidden.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'kohrak_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `We will free Cahdok and Gahdok—and the swarms will be free again. You cannot stop what must be.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `They believe in their mission as fervently as we do in ours. That makes them dangerous.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then we cannot afford to underestimate them. These are not mindless drones—they are elite.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Elite or not, they have what belongs to us. Let's move.`,
+      },
+    ],
+  },
+
+  bohrok_kal_race_to_nest: {
+    id: 'bohrok_kal_race_to_nest',
+    background: { type: 'gradient', from: '#1a0a00', to: '#0d0d1a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The Exo-Toa suits are still down there from our last battle. Without our elements, we'll need every advantage we can get.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I remember those clunkysuits—they held back our powers last time. But without powers to hold back...`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `They have weapons and raw strength. It may be enough.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `It had better be. If the Kal place our symbols on the Nuva Cube and free the Bahrag, the swarm returns—and this time we'll be powerless to stop it.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then there's no time to waste. Downward.`,
+      },
+    ],
+  },
+
+  bohrok_kal_final_confrontation: {
+    id: 'bohrok_kal_final_confrontation',
+    background: { type: 'gradient', from: '#3d0066', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'tahnok_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `You are too late, Toa. The symbols are ours. The queens will be free—and the swarm will cleanse this island as it was always meant to.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'gahlok_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Your Exo-Toa armor is useless against us. Lay down your weapons. This is over.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We have no choice. I must use this.`,
+      },
+      {
+        type: 'narration',
+        text: `Tahu pulls forth the Vahi—the legendary Mask of Time, entrusted to him in secret by Turaga Vakama.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The Vahi! Tahu—do you know what you are doing? If you fail, more than Mata Nui will be lost. This entire reality could unravel!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We have no choice. I must use this power... master it... or all is lost.`,
+      },
+      {
+        type: 'narration',
+        text: `Time slows around the Bohrok-Kal. Their movements crawl to near-stillness. But their Krana-Kal turn silver and project an impenetrable force field. Even frozen in time, they cannot be touched.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I cannot... maintain this... much longer...`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Wait—the symbols! We're still connected to them. If we push our willpower through the connection—feed the Kal more power than they can handle—`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua_Nuva',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Overload them with their own stolen energy. It's brilliant—or insane.`,
+      },
+      {
+        type: 'narration',
+        text: `The six Toa Nuva reach out with their minds, pouring raw willpower into the Nuva Symbols. The Bohrok-Kal convulse as power spirals beyond their control.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'lehvak_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What is happening?! The vacuum—it's pulling outward—I can't contain—`,
+      },
+      {
+        type: 'narration',
+        text: `Lehvak-Kal's vacuum power inverts into a violent thrust. The Kal is launched skyward, rocketing through the cavern ceiling and into the atmosphere—lost to the void above Mata Nui.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'nuhvok_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The gravity... it's folding inward... everything is collapsing... into me...`,
+      },
+      {
+        type: 'narration',
+        text: `Nuhvok-Kal's gravity field spirals out of control. The Kal crumples into itself, crushed into an impossibly dense point—a black hole that swallows its own shell before winking out of existence.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'pahrak_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Too much heat... the plasma... my shell is... melting...`,
+      },
+      {
+        type: 'narration',
+        text: `Pahrak-Kal's plasma burns white-hot, liquefying its own armor. The molten mass sinks through the stone floor, boring deeper and deeper until the glow fades to nothing.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'tahnok_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `No—the electricity—it won't stop arcing—I am... trapped in my own cage...`,
+      },
+      {
+        type: 'narration',
+        text: `Arcs of lightning encase Tahnok-Kal in a crackling prison of its own making. The electricity fuses its joints solid, locking it in place forever—a monument to its own overloaded power.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'gahlok_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The magnetism—everything is pulling toward me—the Exo-Toa debris—I can't repel it—`,
+      },
+      {
+        type: 'narration',
+        text: `Every shard of metal in the cavern hurtles toward Gahlok-Kal. Exo-Toa wreckage, discarded tools, and raw ore slam into the Kal from all directions, burying it under tons of magnetically attracted debris.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'kohrak_kal',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The sonics... the frequency is rising... my own shell is vibrating apart—`,
+      },
+      {
+        type: 'narration',
+        text: `A piercing shriek fills the cavern as Kohrak-Kal's sonic power reaches a frequency beyond endurance. Its armor shatters from within, disintegrating into dust. The silence that follows is absolute.`,
+      },
+      {
+        type: 'narration',
+        text: `The six Nuva Symbols clatter to the ground, unclaimed. The force field dissolves. The Bahrag remain sealed in their protodermis prison.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Our power... I can feel it returning!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The Bahrag remain imprisoned. The Kal are gone. It's over.`,
+      },
+    ],
+  },
+
+  bohrok_kal_naming_day: {
+    id: 'bohrok_kal_naming_day',
+    background: { type: 'gradient', from: '#4a6741', to: '#ffe28e' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `What's happening? Why are the Turaga calling everyone here?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Jala',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `They said it's time for the Matoran to be rebuilt—stronger, taller, better prepared for whatever comes next.`,
+      },
+      {
+        type: 'narration',
+        text: `The Turaga teach the assembled Matoran how to rebuild themselves into sturdier forms. Then, one by one, they call forth heroes to be renamed in honor of their deeds.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Huki',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I am... Hewkii now? It feels different. Stronger.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Maku',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `And I am Macku. A new name for a new era.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Jala',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Jaller. They're calling me Jaller now. Captain Jaller of the Ta-Koro Guard. Has a good ring to it.`,
+      },
+      {
+        type: 'narration',
+        text: `Cheers ring across Kini-Nui as every Matoran is rebuilt and renewed. The island has weathered Makuta, the Bohrok, and the Bohrok-Kal. But the Matoran stand taller now—ready for whatever comes next.`,
+      },
+    ],
+  },
+};
