+++ src/data/cutscenes/bohrok_swarm.ts
@@ -1,0 +1,572 @@
+import type { VisualNovelCutscene } from '../../types/Cutscenes';
+
+export const BOHROK_SWARM_CUTSCENES: Record<string, VisualNovelCutscene> = {
+  bohrok_swarm_intro: {
+    id: 'bohrok_swarm_intro',
+    background: { type: 'gradient', from: '#4a0000', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `What is this?! We just defeated Makuta—the island should be safe!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `These creatures... they're not Rahi. They move with purpose—like a swarm following orders.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I can feel them beneath the ground too. They are coming from below—from deeper than even Makuta's lair.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `They're heading for the villages. All of them.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Le-Koro! My home—the trees are being uptalked—uprooted!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We can't fight them all at once. We need to protect the Matoran first.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `To Ta-Koro—now! We face this new threat together!`,
+      },
+    ],
+  },
+
+  bohrok_ta_koro_defense: {
+    id: 'bohrok_ta_koro_defense',
+    background: { type: 'gradient', from: '#8b0000', to: '#1a0a00' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Tahu! They're breaching the eastern wall!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I see them! Drive them toward the moat—let the lava do the work!`,
+      },
+      {
+        type: 'narration',
+        text: `Tahu channels a wall of flame, herding the Kohrak toward the edge. Several topple into the molten rock, hissing and sparking as they sink.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `It's working! But there are so many...`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then we fight until there are none left. This is our village, Chronicler. We don't yield.`,
+      },
+      {
+        type: 'narration',
+        text: `As the last Kohrak falls, Turaga Vakama emerges from the village. His eyes widen at the sight of the mechanical shells.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `No... the Bohrok. They are real. We prayed the legends were wrong, but the swarms have awakened beneath the island.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `You know what these things are?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I know more than I wish to. Come, Chronicler. We must speak with the Toa—and quickly.`,
+      },
+    ],
+  },
+
+  bohrok_po_koro_defense: {
+    id: 'bohrok_po_koro_defense',
+    background: { type: 'gradient', from: '#c2a366', to: '#4a0000' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Huki',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `They're coming through the pass. Dozens of them. Hafu—we can't hold the gates forever.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Hafu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I know. I have an idea—but you're not going to like it.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Huki',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `What idea?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Hafu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `My statues. The ones lining the canyon approach. If I topple them into the pass, they'll block the tunnel and buy us time to evacuate.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Huki',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Your statues? Hafu, those are your life's work—`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Hafu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `And the villagers are more important. Another Hafu original... this time, a barricade. Stand back.`,
+      },
+      {
+        type: 'narration',
+        text: `With a great heave, Hafu shatters the base of his finest carving. The massive stone figure crashes across the tunnel entrance, blocking the Tahnok's advance. The villagers flee to safety behind the rubble wall.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Huki',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `That was the bravest thing I've ever seen, Hafu.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Hafu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I can always carve more statues. Can't carve more Matoran.`,
+      },
+    ],
+  },
+
+  bohrok_onu_koro_boxor: {
+    id: 'bohrok_onu_koro_boxor',
+    background: { type: 'gradient', from: '#2a1a00', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Onepu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The water's rising. We have maybe an hour before this chamber floods too.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Taipu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Can't we dig through the wall? I'm strong—let me try!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Nuparu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Wait. Look at these parts... the Bohrok shell is hinged. It's designed like a cockpit. If I strip out the Krana control system and rig manual controls...`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Onepu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `You want to build a machine? Now?!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Nuparu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Not just a machine—a fighting vehicle. A Matoran-sized exo-suit. Taipu, hand me that joint assembly. Onepu, I need the pistons from that leg.`,
+      },
+      {
+        type: 'narration',
+        text: `Working frantically as water seeps under the door, Nuparu assembles an ungainly but functional walking machine from Bohrok parts and ancient mining equipment.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Nuparu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I call it... the Boxor. One of you climb in. Let's take back our home.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Taipu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I'll do it! Let me at those Bohrok!`,
+      },
+      {
+        type: 'narration',
+        text: `The Boxor smashes through the flooded tunnel, its mechanical fists scattering Gahlok like leaves. For the first time, the Matoran can fight back.`,
+      },
+    ],
+  },
+
+  bohrok_legend_of_krana: {
+    id: 'bohrok_legend_of_krana',
+    background: { type: 'gradient', from: '#5b3e0b', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I know the answers you seek, though I wish I did not.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What are those creatures, Turaga? They're not like any Rahi we've seen.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `They are called Bohrok. We have known the legends for centuries, and we prayed they were only legends. But the Bohrok are real—all too real. And they are swarming over all of Mata Nui.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What controls them? We pulled a glowing thing from one of their heads.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `That is a Krana—the mind of the swarm. Each Bohrok carries one. There are eight types of Krana, and six breeds of Bohrok. The legends say that only when all forty-eight Krana are collected can the path to the Bohrok nest be unlocked.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `And what's in the nest?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Turaga_Vakama',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The Bahrag—twin queens who command every Bohrok on this island. If the Toa can reach them and defeat them, the swarms will stop. But be warned: the Krana are dangerous. If one attaches to a living being's face... it will control them utterly.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Takua',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `I'll make sure the Toa know. Thank you, Turaga.`,
+      },
+    ],
+  },
+
+  bohrok_lewa_krana_rescue: {
+    id: 'bohrok_lewa_krana_rescue',
+    background: { type: 'gradient', from: '#2d5a1e', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Lewa's mask... he would never leave this behind willingly. Something is very wrong.`,
+      },
+      {
+        type: 'narration',
+        text: `A figure drops from the canopy. It is Lewa—but his movements are rigid, mechanical. A Krana clings to his face where his mask should be. His eyes glow with alien purpose.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The swarm is everything. The swarm will cleanse the island. You cannot stop what must be.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `That's not you talking, brother. Fight it!`,
+      },
+      {
+        type: 'narration',
+        text: `Lewa attacks with devastating wind blasts, but Onua endures. The Toa of Earth is patient and immovable, letting each gust break against him like waves on a cliff.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `You are Lewa—Toa of Air. You love the wind and the trees. You are not a puppet. Remember who you are!`,
+      },
+      {
+        type: 'narration',
+        text: `Onua lunges forward, grabs the Krana, and tears it from Lewa's face in a single motion. Lewa collapses, gasping.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Onua... I could hear myself screaming inside my own head, but my body wouldn't listenheed. Thank you, brother.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `That's what brothers are for. Now—let's make sure this never happens to anyone else.`,
+      },
+    ],
+  },
+
+  bohrok_krana_hunt: {
+    id: 'bohrok_krana_hunt',
+    background: { type: 'gradient', from: '#1a3a4a', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Forty-eight Krana. Eight types, six breeds. That's what we need.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `We should each return to our own region. We know the terrain, and the swarms are concentrated around the villages.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Be careful. Lewa's possession proved these Krana are dangerous. Remove them quickly—don't let them touch your face.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Trust me—I will never let one of those things near me again.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then it's settled. Hunt hard, stay safe, and bring back every last Krana.`,
+      },
+    ],
+  },
+
+  bohrok_into_the_bohrok_nest: {
+    id: 'bohrok_into_the_bohrok_nest',
+    background: { type: 'gradient', from: '#1a0a00', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The tunnel goes deeper than I can see—even with the Akaku. The Bohrok Va have been using this path.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Then this is it. The heart of the swarm is down there.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `I can feel something deep below... something ancient and powerful. Two presences, entwined.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `The Bahrag. The queens of the swarm. If the legends are true, they are what drives every Bohrok on the island.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Lewa',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Last time we went deep-underground, we fought Makuta. This keeps getting better and better.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `At least this time we know what we're getting into.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Together, then. Let's end this.`,
+      },
+    ],
+  },
+
+  bohrok_evolve_toa_nuva: {
+    id: 'bohrok_evolve_toa_nuva',
+    background: { type: 'gradient', from: '#3d0066', to: '#0a0a0a' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Tahu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `The armor is limiting our elemental power! We can't channel through it properly!`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Then we shed the armor. Our elements are what make us Toa—not machines!`,
+      },
+      {
+        type: 'narration',
+        text: `One by one, the Toa discard the Exo-Toa suits. Unshackled, their elemental powers surge—fire, water, ice, air, earth, and stone blazing in concert.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Onua',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Together—now! Trap them!`,
+      },
+      {
+        type: 'narration',
+        text: `Six elemental streams converge on the Bahrag, encasing the twin queens in a prison of solid protodermis. The swarm shudders—and falls silent across the island.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Pohatu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `We did it! They're—wait—the floor!`,
+      },
+      {
+        type: 'narration',
+        text: `The ground cracks open and the six Toa plunge into a pool of silvery, energized protodermis. Power tears through them—reshaping armor, enhancing elements, rewriting their very essence.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Kopaka_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `What... what has happened to us?`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Toa_Gali_Nuva',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `We have changed. I can feel it—our power is greater than before. We are... Toa Nuva.`,
+      },
+    ],
+  },
+
+  bohrok_assistants: {
+    id: 'bohrok_assistants',
+    background: { type: 'gradient', from: '#2a1a00', to: '#1a2a00' },
+    steps: [
+      {
+        type: 'dialogue',
+        speakerId: 'Nuparu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `Onepu, look at these Bohrok shells. The Krana are inert, but the mechanical systems are intact. If I remove the Krana and install manual control rigs—like the Boxor, but simpler—we could repurpose them.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Onepu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `Repurpose the things that nearly destroyed our homes? The other Matoran won't like it.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Nuparu',
+        portraitType: 'avatar',
+        position: 'left',
+        text: `They'll like the free labor. These shells can haul stone, clear rubble, dig tunnels—everything we need for reconstruction. We turn our enemy's weapons into tools.`,
+      },
+      {
+        type: 'dialogue',
+        speakerId: 'Onepu',
+        portraitType: 'avatar',
+        position: 'right',
+        text: `You're mad. But you were mad when you built the Boxor too, and that saved us all. Fine—let's get to work.`,
+      },
+    ],
+  },
+};
